Break down test_slurm_command_gen_strategy into smaller tests
