+++ tests/slurm_command_gen_strategy/test_nccl.py
@@ -1,0 +1,51 @@
+# SPDX-FileCopyrightText: NVIDIA CORPORATION & AFFILIATES
+# Copyright (c) 2024 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+
+from cloudai.schema.test_template.nccl_test.slurm_command_gen_strategy import NcclTestSlurmCommandGenStrategy
+from cloudai.systems import SlurmSystem
+
+
+class TestNCCLSlurmCommandGen:
+    def get_cmd(self, slurm_system: SlurmSystem, slurm_args: dict, cmd_args: dict) -> str:
+        return NcclTestSlurmCommandGenStrategy(slurm_system, {}).generate_srun_command(slurm_args, {}, cmd_args, "")
+
+    def test_only_mandatory(self, slurm_system: SlurmSystem) -> None:
+        slurm_args = {"image_path": "fake_image_path"}
+        cmd_args = {"subtest_name": "fake_subtest_name"}
+        cmd = self.get_cmd(slurm_system, slurm_args, cmd_args)
+        assert cmd == " \\\n".join(
+            [
+                "srun",
+                f"--mpi={slurm_system.mpi}",
+                f"--container-image={slurm_args['image_path']}",
+                f"/usr/local/bin/{cmd_args['subtest_name']}",
+            ]
+        )
+
+    def test_with_container_mounts(self, slurm_system: SlurmSystem) -> None:
+        slurm_args = {"image_path": "fake_image_path", "container_mounts": "fake_mounts"}
+        cmd_args = {"subtest_name": "fake_subtest_name"}
+        cmd = self.get_cmd(slurm_system, slurm_args, cmd_args)
+        assert cmd == " \\\n".join(
+            [
+                "srun",
+                f"--mpi={slurm_system.mpi}",
+                f"--container-image={slurm_args['image_path']}",
+                f"--container-mounts={slurm_args['container_mounts']}",
+                f"/usr/local/bin/{cmd_args['subtest_name']}",
+            ]
+        )
